Adding kafka utils
