+++ koku/kafka_utils/utils.py
@@ -8,12 +8,36 @@ import random
 import socket
 import time
 
+from confluent_kafka import Consumer
+from confluent_kafka import Producer
 from kafka import BrokerConnection
 
+from masu.config import Config
 from masu.prometheus_stats import KAFKA_CONNECTION_ERRORS_COUNTER
 
-
 LOG = logging.getLogger(__name__)
+
+
+def get_consumer(topic, address):  # pragma: no cover
+    """Create a Kafka consumer."""
+    consumer = Consumer(
+        {
+            "bootstrap.servers": address,
+            "group.id": "hccm-group",
+            "queued.max.messages.kbytes": 1024,
+            "enable.auto.commit": False,
+            "max.poll.interval.ms": 1080000,  # 18 minutes
+        },
+        logger=LOG,
+    )
+    consumer.subscribe([topic])
+    return consumer
+
+
+def get_producer():  # pragma: no cover
+    """Create a Kafka producer."""
+    producer = Producer({"bootstrap.servers": Config.INSIGHTS_KAFKA_ADDRESS, "message.timeout.ms": 1000})
+    return producer
 
 
 def backoff(interval, maximum=120):
