+++ koku/masu/external/kafka_msg_handler.py
@@ -13,7 +13,6 @@ import tempfile
 import threading
 import time
 import traceback
-import uuid
 from tarfile import ReadError
 from tarfile import TarFile
 
@@ -34,7 +33,6 @@ from masu.database.report_manifest_db_accessor import ReportManifestDBAccessor
 from masu.external import UNCOMPRESSED
 from masu.external.accounts_accessor import AccountsAccessor
 from masu.external.accounts_accessor import AccountsAccessorError
-from masu.external.date_accessor import DateAccessor
 from masu.external.downloader.ocp.ocp_report_downloader import create_daily_archives
 from masu.external.downloader.ocp.ocp_report_downloader import OCPReportDownloader
 from masu.processor._tasks.process import _process_report_file
@@ -794,48 +792,3 @@ def initialize_kafka_handler():  # pragma: no cover
         event_loop_thread.daemon = True
         event_loop_thread.start()
         event_loop_thread.join()
-
-
-@KAFKA_CONNECTION_ERRORS_COUNTER.count_exceptions()
-def send_notification(account, event_type):
-    """
-    Send kafka notification message to Insights Notifications Service.
-    Args:
-        account (Object): account for notifications being sent.
-        notification (String): Notification string
-        uuid (UUID): uuid of object in question
-    Returns:
-        None
-    """
-    encoded_id = str(uuid.uuid4()).encode()
-    message_id = bytearray(encoded_id)
-    producer = get_producer()
-    notification_json = {
-        "id": message_id,
-        "bundle": "openshift",
-        "application": "cost-management",
-        "event_type": event_type,
-        "timestamp": DateAccessor().today(),
-        "account_id": account.get("schema_name"),
-        "org_id": account.get("org_id"),
-        "context": {
-            "source_id": account.get("provider_uuid"),
-            "source_name": account.get("name"),
-            "host_url": f"https://console.redhat.com/settings/sources/detail/{account.get('provider_uuid')}",
-        },
-        "events": [
-            {
-                "metadata": {},
-                "payload": {
-                    "description": "Openshift source has no cost model assigned, add one via the following link.",
-                    "host_url": "https://console.redhat.com/openshift/cost-management/cost-models",
-                },
-            }
-        ],
-    }
-    msg = bytes(json.dumps(notification_json), "utf-8")
-    producer.produce(Config.NOTIFICATION_TOPIC, value=msg, callback=delivery_callback)
-    # Wait up to 1 second for events. Callbacks will be invoked during
-    # this method call if the message is acknowledged.
-    # `flush` makes this process synchronous compared to async with `poll`
-    producer.flush(1)
